pyproject.toml
[tool.poetry]
name = "kurs7"
version = "0.1.0"
description = ""
authors = ["Aidar Putilov <putiloff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
requests = "^2.32.3"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
coverage = "^7.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
